refactor(journal$): gather related functions under common module
