--- src/journal/createJournal$.type.ts
@@ -1,4 +1,0 @@
-import {Observable} from "rxjs";
-import {Journal} from "./journal.type";
-
-export type CreateJournal$ = (userId: string, name: string) => Observable<Journal>;
