+++ src/journal/journal$.type.ts
@@ -2,3 +2,13 @@ import {Observable} from "rxjs";
 import {Journal} from "./journal.type";
 
 export type Journal$ = (id: string) => Observable<Journal | undefined>;
+export type Journals$ = (id: string, page: number, limit: number) => Observable<Journal[]>;
+export type CreateJournal$ = (userId: string, name: string) => Observable<Journal>;
+export type DeleteJournal$ = (id: string) => Observable<Journal>;
+export type UpdateJournal$ = (id: string, name: string) => Observable<Journal>;
+export type SortUsersJournals$ = (
+    id: string,
+    order: 1 | -1,
+    page: number,
+    limit: number,
+) => Observable<Journal[]>;
