--- src/journal/journals$.type.ts
@@ -1,5 +1,0 @@
-import {Observable} from "rxjs";
-import {Journal} from "./journal.type";
-
-export type Journals$ =
-    (id: string, page: number, limit: number) => Observable<Journal[]>;
