--- src/journal/updateJournal$.type.ts
@@ -1,4 +1,0 @@
-import {Observable} from "rxjs";
-import {Journal} from "../journal/journal.type";
-
-export type UpdateJournal$ = (id: string, name: string) => Observable<Journal>;
